+++ monkey/common/hard_coded_manifests/hard_coded_credentials_collector_manifests.py
@@ -1,10 +1,10 @@
 from common.agent_plugins import AgentPluginManifest, AgentPluginType
 from common.operating_system import OperatingSystem
 
-HARD_CODED_CREDENTIAL_COLLECTOR_MANIFESTS = {
+HARD_CODED_CREDENTIALS_COLLECTOR_MANIFESTS = {
     "MimikatzCollector": AgentPluginManifest(
         name="MimikatzCollector",
-        plugin_type=AgentPluginType.CREDENTIAL_COLLECTOR,
+        plugin_type=AgentPluginType.CREDENTIALS_COLLECTOR,
         supported_operating_systems=(OperatingSystem.LINUX, OperatingSystem.WINDOWS),
         target_operating_systems=(OperatingSystem.WINDOWS,),
         title="Mimikatz Credentials Collector",
@@ -14,7 +14,7 @@ HARD_CODED_CREDENTIAL_COLLECTOR_MANIFESTS = {
     ),
     "SSHCollector": AgentPluginManifest(
         name="SSHCollector",
-        plugin_type=AgentPluginType.CREDENTIAL_COLLECTOR,
+        plugin_type=AgentPluginType.CREDENTIALS_COLLECTOR,
         supported_operating_systems=(OperatingSystem.LINUX, OperatingSystem.WINDOWS),
         target_operating_systems=(OperatingSystem.LINUX,),
         title="SSH Credentials Collector",