+++ monkey/monkey_island/cc/services/agent_configuration_service/hard_coded_schemas/hard_coded_credentials_collector_schemas.py
@@ -1,4 +1,4 @@
-HARD_CODED_CREDENTIAL_COLLECTOR_SCHEMAS = {
+HARD_CODED_CREDENTIALS_COLLECTOR_SCHEMAS = {
     "MimikatzCollector": {"type": "object", "properties": {}},
     "SSHCollector": {"type": "object", "properties": {}},
 }