+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - `PortScanData.open` property. #3238
 
+### Changed
+- Renamed "Credential collector" to "Credentials collector". #3167
+
 ## [2.1.0] - 2023-04-19
 ### Added
 - Logout button. #3063